Remove PublishFolders task and update README
Initially I wanted to update it to use the same code as PublishFiles and remove the redundant WithHeaders task, but there were two major issues. The biggest is to support recursive folder publishing (preserving the sub-folder structure) would require a very large overhaul of the entire internal API (while it's easy to implement by calling with MSBuild ItemGroups). In addition, PublishFolderWithHeaders uses a property SourceFolders while PublishFolder uses SourceFolder which is a pain.